Different map check method
This one doesn't seem to give me the same issues, but I'm worried any attempt to typecheck the map will run into file ordering issues, which I assume is what's causing these heisenbugs
Linting

My bad, didn't realize linting was on, I'll check in the future
Disable linting for map checker

It's not a good idea to have a linter check frankencode like that anyway. Liable to break in the future if you do.