Add Ruby dependencies for local docs development
It's unfortunate to have to require Ruby as a dependency in this project, but in order to work with the documentation and extend it for GitHub Pages this makes it much easier to contribute because changes can be previewed locally.

To sort of keep the Ruby dependency contained, we add this to the /docs folder, but there would be nothing stopping us from adding this to the root, which would be a more usual place to look for these files.