Fix broken documentation links
I ran `npx broken-link-checker -ro https://gcanti.github.io/fp-ts/` to discover more broken links. Ideally, I would like to test for broken links with `npm test`, but I'm not sure this is feasible because it makes most sense to run it on the deployed site, but we don't have that at time of testing. It's also slow. Let's see what the future holds …

Closes #791, #792