doc: add type signature for `TaskEither` in module intro
Thanks for this excellent and remarkably complete library!

This adds the type signature of `TaskEither` to the introductory
paragraph on the `TaskEither` page in a way that aligns with the `Task`
docs, which also have the signature at the top.

I'm opening this because when I first used `TaskEither`, I hadn't used
`Task` yet, either, and I didn't realize how to get the promise out of
it. I started out by looking for some kind of destructor method rather
than thinking it was callable. It was only when I looked at the `Task`
documentation that I realized it might be callable.

It's possible that if the top of the file had directed me that it was
an extension of `Task`, I might have figured it out sooner. I should
have thought to look down around line 860, where the existing docs
describe the signature, but I did not.